refactor: move /internal pkgs to /testutil
