+++ pkg/observable/channel/observable_test.go
@@ -10,10 +10,10 @@ import (
 	"github.com/stretchr/testify/require"
 	"golang.org/x/sync/errgroup"
 
-	"github.com/pokt-network/poktroll/internal/testchannel"
-	"github.com/pokt-network/poktroll/internal/testerrors"
 	"github.com/pokt-network/poktroll/pkg/observable"
 	"github.com/pokt-network/poktroll/pkg/observable/channel"
+	"github.com/pokt-network/poktroll/testutil/testchannel"
+	"github.com/pokt-network/poktroll/testutil/testerrors"
 )
 
 const (
