+++ testutil/testchannel/drain.go
